+++ .travis.yml
@@ -2,7 +2,6 @@ language: 'php'
 php:
  - '7.3'
 before_install: true
-install: 
  - composer install
 
 before_script: true
